// client/src/lib/shared/bindings/GameStateServer.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Board } from "./Board";

export type GameStateServer = { white: Board, black: Board, white_player_id: string, black_player_id: string, white_turn: boolean, turn_count: number, cards_played_this_turn: number, };