// client/src/lib/shared/bindings/IncantationEntity.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { IncantationCard } from "./IncantationCard";

export type IncantationEntity = { entity_id: string, cost: number, turns_in_hand: number, just_drawn: boolean, card: IncantationCard, };