// client/src/lib/shared/bindings/TargetFilter.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MinionAttribute } from "./MinionAttribute";
import type { Race } from "./Race";

export type TargetFilter = { "IsRace": { race: Race, } } | { "HasAttribute": { attribute: MinionAttribute, } };