// client/src/lib/shared/bindings/TargetSpec.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntityType } from "./EntityType";
import type { TargetFilter } from "./TargetFilter";
import type { TargetMode } from "./TargetMode";
import type { TargetSide } from "./TargetSide";

export type TargetSpec = { target_mode: TargetMode, side: TargetSide, entity_type: EntityType, filters: Array<TargetFilter>, };